Refactored IT into seperate files
